Caluromys
The woolly opossum of the genus Caluromys, despite their wide distribution range, are poorly known species.  In Colombia, the presence of C. derbianus and C. lanatus is recognized, and the presence of C. philander has been proposed.  This study sought to update the distribution of the species of the genus Caluromys in Colombia.  We modeled the current and potential distribution of C. derbianus and C. lanatus and discussed the potential presence of C. philander.  We extend the distribution range for C. derbianus and C. lanatus by obtaining new records mainly in the Colombian Caribbean region.  For C. derbianus, we found that the estimated model area was 177,337 km2, with the most suitable areas in the Caribbean region in the department of Córdoba and the coast of the department of Sucre, in the north and south of the Pacific region, and in the Andean region on the Western Mountain Range, associated with forest cover.  For C. lanatus, the estimated model area was 940,007 km2 with the greatest habitat suitability in the forest cover of the Sierra Nevada de Santa Marta, the Central Mountain Range, the Magdalena river valleys and the Andean-Orinosence piedmont.  The areas of occupancy (AOO) and extent of occurrence (EOO) in Colombia are 268 km2 and 182,741 km2 for C. derbianus, 652 km2 and 1,036,486 km2 for C. lanatus.  Although we found high uncertainty in the records for C. philander in Colombia, we consider that the species probably has representation in the eastern Amazon and Orinoquia regions.  This information could be useful for assessing threats to woolly opossums in Colombia.